Arrumado banco de dados, excluido tabela compra, criado medoto de pesquisar comanda por id, editar desconto e outro

+++ src/main/kotlin/br/com/sapiencia/command/database/entity/Pagamento.kt
@@ -1,31 +1,28 @@
 package br.com.sapiencia.command.database.entity
 
-import br.com.sapiencia.command.model.MesaModel
 import br.com.sapiencia.command.model.PagamentoModel
-import br.com.sapiencia.command.model.ProdutoModel
-import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
-import jakarta.persistence.OneToOne
 import jakarta.persistence.Table
 import org.hibernate.annotations.GenericGenerator
+import java.math.BigDecimal
 import java.util.UUID
 
 @Entity
 @Table(name = "pagamento")
-data class Pagamento (
+data class Pagamento(
     @Id
     @GeneratedValue(generator = "uuid-hibernate-generator")
     @GenericGenerator(name = "uuid-hibernate-generator", strategy = "org.hibernate.id.UUIDGenerator")
     val id: UUID? = null,
-    val valorPago:Double,
+    val valorPago: BigDecimal,
     @ManyToOne
     @JoinColumn(name = "meio_pagamento_id")
-    val meioPagamento:MeioPagamento
-){
+    val meioPagamento: MeioPagamento
+) {
     fun toModel() = PagamentoModel(
         id = id,
         valorPago = valorPago,
